promote unilateral-exit test helpers to spark-itest::helpers
Move FundedUtxo, fund_p2tr_utxo, fund_p2wpkh_utxo, make_cpfp_input, the
sign_cpfp_psbt_* signers, and submit_package_with_csv_retry from the
local_unilateral_exit_tests test file into the library's helpers module so
they can be reused from breez-itest's unilateral exit tests.

Adds a pub `rpc` accessor on BitcoindFixture to let test-only chain-service
adapters reach arbitrary bitcoind methods without exposing the full
rpc_call impl.
