+++ crates/spark-itest/tests/local_unilateral_exit_tests.rs
@@ -2,227 +2,21 @@ use std::collections::HashSet;
 
 use anyhow::{Result, bail};
 use bitcoin::{
-    Address, Amount, CompressedPublicKey, OutPoint, Psbt, Sequence, Transaction, TxIn, TxOut, Txid,
-    Witness,
+    Amount, OutPoint, Psbt, Sequence, Transaction, TxIn, TxOut, Txid, Witness,
     absolute::LockTime,
     ecdsa::Signature as EcdsaSignature,
     hashes::Hash as _,
     key::{Secp256k1, TapTweak as _},
-    secp256k1::SecretKey,
     sighash::{self, SighashCache},
     transaction::Version,
 };
 use rstest::*;
-use spark_itest::{
-    fixtures::bitcoind::BitcoindFixture,
-    helpers::{WalletsFixture, deposit_to_wallet, deposit_with_amount, wallets},
+use spark_itest::helpers::{
+    P2TR_INPUT_WEIGHT, P2WPKH_INPUT_WEIGHT, WalletsFixture, deposit_to_wallet, deposit_with_amount,
+    fund_p2tr_utxo, fund_p2wpkh_utxo, make_cpfp_input, sign_cpfp_psbt_custom, sign_cpfp_psbt_p2tr,
+    sign_cpfp_psbt_p2wpkh, wallets,
 };
-use spark_wallet::CpfpInput;
 use tracing::info;
-
-// ---------------------------------------------------------------------------
-// Constants
-// ---------------------------------------------------------------------------
-
-/// P2TR key-path signed input weight: 41 non-witness × 4 + 66 witness = 230 WU
-const P2TR_INPUT_WEIGHT: u64 = 230;
-/// P2WPKH signed input weight: 41 non-witness × 4 + 108 witness = 272 WU
-const P2WPKH_INPUT_WEIGHT: u64 = 272;
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-/// Funded UTXO ready for use as a CPFP input.
-struct FundedUtxo {
-    secret_key: SecretKey,
-    outpoint: OutPoint,
-    witness_utxo: TxOut,
-    address: Address,
-}
-
-/// Fund a new P2TR address from bitcoind and return the UTXO details.
-async fn fund_p2tr_utxo(bitcoind: &BitcoindFixture, amount: Amount) -> Result<FundedUtxo> {
-    let secp = Secp256k1::new();
-    let secret_key = SecretKey::new(&mut rand::thread_rng());
-    let pubkey = secret_key.public_key(&secp);
-    let (xonly, _) = pubkey.x_only_public_key();
-    let address = Address::p2tr(&secp, xonly, None, bitcoin::Network::Regtest);
-
-    let txid = bitcoind.fund_address(&address, amount).await?;
-    bitcoind.generate_blocks(1).await?;
-    bitcoind.wait_for_tx_confirmation(&txid, 1).await?;
-
-    let tx = bitcoind.get_transaction(&txid).await?;
-    let vout = tx
-        .output
-        .iter()
-        .position(|o| o.script_pubkey == address.script_pubkey())
-        .expect("P2TR output not found") as u32;
-
-    Ok(FundedUtxo {
-        secret_key,
-        outpoint: OutPoint { txid, vout },
-        witness_utxo: TxOut {
-            value: amount,
-            script_pubkey: address.script_pubkey(),
-        },
-        address,
-    })
-}
-
-/// Fund a new P2WPKH address from bitcoind and return the UTXO details.
-async fn fund_p2wpkh_utxo(bitcoind: &BitcoindFixture, amount: Amount) -> Result<FundedUtxo> {
-    let secp = Secp256k1::new();
-    let secret_key = SecretKey::new(&mut rand::thread_rng());
-    let pubkey = secret_key.public_key(&secp);
-    let compressed = CompressedPublicKey(pubkey);
-    let address = Address::p2wpkh(&compressed, bitcoin::Network::Regtest);
-
-    let txid = bitcoind.fund_address(&address, amount).await?;
-    bitcoind.generate_blocks(1).await?;
-    bitcoind.wait_for_tx_confirmation(&txid, 1).await?;
-
-    let tx = bitcoind.get_transaction(&txid).await?;
-    let vout = tx
-        .output
-        .iter()
-        .position(|o| o.script_pubkey == address.script_pubkey())
-        .expect("P2WPKH output not found") as u32;
-
-    Ok(FundedUtxo {
-        secret_key,
-        outpoint: OutPoint { txid, vout },
-        witness_utxo: TxOut {
-            value: amount,
-            script_pubkey: address.script_pubkey(),
-        },
-        address,
-    })
-}
-
-fn make_cpfp_input(utxo: &FundedUtxo, weight: u64) -> CpfpInput {
-    CpfpInput {
-        outpoint: utxo.outpoint,
-        witness_utxo: utxo.witness_utxo.clone(),
-        signed_input_weight: weight,
-    }
-}
-
-/// Sign a CPFP PSBT that has P2TR external inputs. Finalizes anchor + P2TR inputs.
-fn sign_cpfp_psbt_p2tr(psbt: &Psbt, secret_key: &SecretKey) -> Result<Transaction> {
-    let mut psbt = psbt.clone();
-    finalize_anchor_inputs(&mut psbt);
-
-    let secp = Secp256k1::new();
-    let keypair = bitcoin::key::Keypair::from_secret_key(&secp, secret_key)
-        .tap_tweak(&secp, None)
-        .to_keypair();
-
-    let prevouts: Vec<TxOut> = psbt
-        .inputs
-        .iter()
-        .map(|i| i.witness_utxo.clone().unwrap_or(TxOut::NULL))
-        .collect();
-    let prevouts_ref = sighash::Prevouts::All(&prevouts);
-
-    let taproot_indices: Vec<usize> = psbt
-        .inputs
-        .iter()
-        .enumerate()
-        .filter(|(_, i)| {
-            i.final_script_witness.is_none()
-                && i.witness_utxo
-                    .as_ref()
-                    .is_some_and(|o| o.script_pubkey.is_p2tr())
-        })
-        .map(|(idx, _)| idx)
-        .collect();
-
-    let mut cache = SighashCache::new(&psbt.unsigned_tx);
-    for i in taproot_indices {
-        let sighash = cache.taproot_key_spend_signature_hash(
-            i,
-            &prevouts_ref,
-            sighash::TapSighashType::Default,
-        )?;
-        let msg = bitcoin::secp256k1::Message::from_digest(sighash.to_byte_array());
-        let schnorr_sig = secp.sign_schnorr_no_aux_rand(&msg, &keypair);
-        let tap_sig = bitcoin::taproot::Signature {
-            signature: schnorr_sig,
-            sighash_type: sighash::TapSighashType::Default,
-        };
-        let mut witness = Witness::new();
-        witness.push(tap_sig.to_vec());
-        psbt.inputs[i].final_script_witness = Some(witness);
-    }
-
-    Ok(psbt.extract_tx_unchecked_fee_rate())
-}
-
-/// Sign a CPFP PSBT that has P2WPKH external inputs. Finalizes anchor + P2WPKH inputs.
-fn sign_cpfp_psbt_p2wpkh(psbt: &Psbt, secret_key: &SecretKey) -> Result<Transaction> {
-    let mut psbt = psbt.clone();
-    finalize_anchor_inputs(&mut psbt);
-
-    let secp = Secp256k1::new();
-    let pubkey = secret_key.public_key(&secp);
-    let bitcoin_pubkey = bitcoin::PublicKey::new(pubkey);
-
-    let wpkh_indices: Vec<usize> = psbt
-        .inputs
-        .iter()
-        .enumerate()
-        .filter(|(_, i)| {
-            i.final_script_witness.is_none()
-                && i.witness_utxo
-                    .as_ref()
-                    .is_some_and(|o| o.script_pubkey.is_p2wpkh())
-        })
-        .map(|(idx, _)| idx)
-        .collect();
-
-    let mut cache = SighashCache::new(&psbt.unsigned_tx);
-    for i in wpkh_indices {
-        let (msg, ecdsa_type) = psbt
-            .sighash_ecdsa(i, &mut cache)
-            .map_err(|e| anyhow::anyhow!("ECDSA sighash error: {e}"))?;
-        let sig = secp.sign_ecdsa(&msg, secret_key);
-        let signature = EcdsaSignature {
-            signature: sig,
-            sighash_type: ecdsa_type,
-        };
-        let mut witness = Witness::new();
-        witness.push(signature.to_vec());
-        witness.push(bitcoin_pubkey.to_bytes());
-        psbt.inputs[i].final_script_witness = Some(witness);
-    }
-
-    Ok(psbt.extract_tx_unchecked_fee_rate())
-}
-
-/// Sign a CPFP PSBT using a caller-provided closure (custom signer).
-fn sign_cpfp_psbt_custom<F>(psbt: &Psbt, signer: F) -> Result<Transaction>
-where
-    F: FnOnce(&mut Psbt) -> Result<()>,
-{
-    let mut psbt = psbt.clone();
-    finalize_anchor_inputs(&mut psbt);
-    signer(&mut psbt)?;
-    Ok(psbt.extract_tx_unchecked_fee_rate())
-}
-
-/// Finalize all ephemeral anchor inputs in a PSBT with an empty witness.
-fn finalize_anchor_inputs(psbt: &mut Psbt) {
-    for input in &mut psbt.inputs {
-        if let Some(ref tx_out) = input.witness_utxo
-            && tx_out.value.to_sat() == 0
-            && tx_out.script_pubkey.as_bytes() == [0x51, 0x02, 0x4e, 0x73]
-        {
-            input.final_script_witness = Some(Witness::new());
-        }
-    }
-}
 
 // ---------------------------------------------------------------------------
 // Standalone Signature Tests
